+++ bioshell-cartesians/src/coordinates.rs
@@ -2,11 +2,7 @@ use std::ops::{Index, IndexMut, Range};
 
 use bioshell_numerical::Vec3;
 use bioshell_sim::{ResizableSystem, System};
-
-/// Stateless immutable view of coordinates
-pub struct CoordinatesView<'a> {
-    pub points: &'a Coordinates,
-}
+use crate::{closest_image, wrap_coordinate_to_box};
 
 /// Represents a system of Vec3 points in a simulation.
 ///
@@ -14,50 +10,9 @@ pub struct CoordinatesView<'a> {
 #[derive(Clone, Debug)]
 pub struct Coordinates {
     box_len: f64,
-    box_len_half: f64,
     current_size: usize,
     coords_vec: Vec<Vec3>,
     chains: Vec<Range<usize>>,
-}
-
-macro_rules! wrap_coordinate_to_box {
-    ($val:expr, $L:expr, $coord:expr) => {
-        $coord = $val;
-        if $coord > $L {
-            $coord = $coord - $L
-        } else {
-            if $coord < 0.0 {
-                $coord = $L + $coord
-            }
-        }
-    };
-}
-
-/// Calculates the shortest difference `c1 - c2`, taking periodic boundary condition into account
-macro_rules! closest_image {
-    ($c1:expr, $c2:expr, $L: expr,$L2: expr, $delta:expr) => {
-        $delta = $c1 - $c2;
-        if $delta > 0.0 {
-            if $delta > $L2 {
-                $delta -= $L
-            }
-        } else {
-            if $delta < -$L2 {
-                $delta += $L
-            }
-        }
-    };
-}
-
-/// Finds the length of a simulation box to achieve assumed density
-///
-/// # Arguments
-/// * `atom_radius` - atomic radius is same for all atoms of the system
-/// * `n_atoms` - number of atoms in the system
-/// * `density` - target density
-pub fn box_width(atom_radius: f64, n_atoms: usize, density: f64) -> f64 {
-    let v: f64 = 4.0 / 3.0 * std::f64::consts::PI * atom_radius.powi(3);
-    (n_atoms as f64 * v / density).powf(1.0 / 3.0)
 }
 
 impl Coordinates {
@@ -82,25 +37,27 @@ impl Coordinates {
         return Coordinates {
             current_size: 0,
             box_len: l,
-            box_len_half: l / 2.0,
             coords_vec: v,
             chains,
         };
     }
 
     #[inline(always)]
-    pub fn box_len(&self) -> f64 {
-        self.box_len
+    pub fn get_box_len(&self) -> f64 {
+        return self.box_len;
+    }
+
+    pub fn get_box_len_half(&self)->f64{
+        return self.get_box_len()/2.0;
     }
 
     #[inline(always)]
     pub fn set_box_len(&mut self, new_box_len: f64) {
         self.box_len = new_box_len;
-        self.box_len_half = new_box_len / 2.0;
     }
 
     /// Returns the number of chains in this system
-    pub fn count_chains(&self) -> usize {
+    pub fn get_chains_count(&self) -> usize {
         return self.chains.len();
     }
 
@@ -108,7 +65,7 @@ impl Coordinates {
     ///
     /// Per rust convention used in ``std::ops::Range`` struct, the returned
     /// ``start..end`` range contains all atoms indexed by ``start <= idx < end``
-    pub fn chain_range(&self, idx: usize) -> &Range<usize> {
+    pub fn get_chain_range(&self, idx: usize) -> &Range<usize> {
         &self.chains[idx]
     }
 
@@ -117,7 +74,7 @@ impl Coordinates {
     /// # Arguments
     /// * `i` - index of the first atom
     /// * `j` - index of the second atom
-    pub fn distance_square(&self, i: usize, j: usize) -> f64 {
+    pub fn get_distance_square(&self, i: usize, j: usize) -> f64 {
         let mut d = self.coords_vec[i].x - self.coords_vec[j].x;
         let mut d2 = d * d;
         d = self.coords_vec[i].y - self.coords_vec[j].y;
@@ -134,13 +91,13 @@ impl Coordinates {
     /// # Arguments
     /// * `i` - index of the first atom
     /// * `j` - index of the second atom
-    pub fn closest_distance_square(&self, i: usize, j: usize) -> f64 {
+    pub fn get_closest_distance_square(&self, i: usize, j: usize) -> f64 {
         let mut d: f64;
-        closest_image!(self.coords_vec[i].x, self.coords_vec[j].x, self.box_len, self.box_len_half, d);
+        closest_image!(self.coords_vec[i].x, self.coords_vec[j].x, self.box_len, self.get_box_len_half(), d);
         let mut d2 = d * d;
-        closest_image!(self.coords_vec[i].y, self.coords_vec[j].y, self.box_len, self.box_len_half, d);
+        closest_image!(self.coords_vec[i].y, self.coords_vec[j].y, self.box_len, self.get_box_len_half(), d);
         d2 += d * d;
-        closest_image!(self.coords_vec[i].z, self.coords_vec[j].z, self.box_len, self.box_len_half, d);
+        closest_image!(self.coords_vec[i].z, self.coords_vec[j].z, self.box_len, self.get_box_len_half(), d);
 
         return d2 + d * d;
     }
@@ -152,13 +109,13 @@ impl Coordinates {
     /// # Arguments
     /// * `i` - index of the first atom
     /// * `v` - position of the second atom
-    pub fn closest_distance_square_to_vec(&self, i: usize, v: &Vec3) -> f64 {
+    pub fn get_closest_distance_square_to_vec(&self, i: usize, v: &Vec3) -> f64 {
         let mut d: f64;
-        closest_image!(self.coords_vec[i].x, v.x, self.box_len, self.box_len_half, d);
+        closest_image!(self.coords_vec[i].x, v.x, self.box_len, self.get_box_len_half(), d);
         let mut d2 = d * d;
-        closest_image!(self.coords_vec[i].y, v.y, self.box_len, self.box_len_half, d);
+        closest_image!(self.coords_vec[i].y, v.y, self.box_len, self.get_box_len_half(), d);
         d2 += d * d;
-        closest_image!(self.coords_vec[i].z, v.z, self.box_len, self.box_len_half, d);
+        closest_image!(self.coords_vec[i].z, v.z, self.box_len, self.get_box_len_half(), d);
 
         return d2 + d * d;
     }
@@ -168,15 +125,15 @@ impl Coordinates {
     /// # Arguments
     /// * `ref_atom` - the reference atom
     /// * `the_atom` - index of the atom to be cloned
-    pub fn clone_closest_image(&self, ref_atom: usize, the_atom: usize) -> Vec3 {
+    pub fn get_closest_image_clone(&self, ref_atom: usize, the_atom: usize) -> Vec3 {
 
         let mut out = self.coords_vec[ref_atom].clone();
         let mut d: f64;
-        closest_image!(self.coords_vec[the_atom].x, self.coords_vec[ref_atom].x, self.box_len, self.box_len_half, d);
+        closest_image!(self.coords_vec[the_atom].x, self.coords_vec[ref_atom].x, self.box_len, self.get_box_len_half(), d);
         out.x += d;
-        closest_image!(self.coords_vec[the_atom].y, self.coords_vec[ref_atom].y, self.box_len, self.box_len_half, d);
+        closest_image!(self.coords_vec[the_atom].y, self.coords_vec[ref_atom].y, self.box_len, self.get_box_len_half(), d);
         out.y += d;
-        closest_image!(self.coords_vec[the_atom].z, self.coords_vec[ref_atom].z, self.box_len, self.box_len_half, d);
+        closest_image!(self.coords_vec[the_atom].z, self.coords_vec[ref_atom].z, self.box_len, self.get_box_len_half(), d);
         out.z += d;
 
         return out;
@@ -185,42 +142,42 @@ impl Coordinates {
     /// Calculates the difference in ``x`` coordinate between the i-th atom and a given ``x`` value
     /// This function obeys periodic boundary conditions and returns the distance to the closest
     /// image of the  position ``i``
-    pub fn delta_x(&self, i: usize, x: f64) -> f64 {
+    pub fn get_delta_x(&self, i: usize, x: f64) -> f64 {
         let mut d: f64;
-        closest_image!(self.coords_vec[i].x, x, self.box_len, self.box_len_half, d);
+        closest_image!(self.coords_vec[i].x, x, self.box_len, self.get_box_len_half(), d);
         d
     }
 
     /// Calculates the difference in ``y`` coordinate between the i-th atom and a given ``y`` value
     /// This function obeys periodic boundary conditions and returns the distance to the closest
     /// image of the  position ``i``
-    pub fn delta_y(&self, i: usize, y: f64) -> f64 {
+    pub fn get_delta_y(&self, i: usize, y: f64) -> f64 {
         let mut d: f64;
-        closest_image!(self.coords_vec[i].y, y, self.box_len, self.box_len_half, d);
+        closest_image!(self.coords_vec[i].y, y, self.box_len, self.get_box_len_half(), d);
         d
     }
 
     /// Calculates the difference in ``z`` coordinate between the i-th atom and a given ``z`` value
     /// This function obeys periodic boundary conditions and returns the distance to the closest
     /// image of the  position ``i``
-    pub fn delta_z(&self, i: usize, z: f64) -> f64 {
+    pub fn get_delta_z(&self, i: usize, z: f64) -> f64 {
         let mut d: f64;
-        closest_image!(self.coords_vec[i].z, z, self.box_len, self.box_len_half, d);
+        closest_image!(self.coords_vec[i].z, z, self.box_len, self.get_box_len_half(), d);
         d
     }
 
     /// Provides the chain index type of the `i`-th atom of this system
-    pub fn chain_id(&self, i: usize) -> u16 {
+    pub fn get_chain_id(&self, i: usize) -> u16 {
         self.coords_vec[i].chain_id
     }
 
     /// Provides the residue type of the `i`-th **atom** of this system
-    pub fn res_type(&self, i: usize) -> u8 {
+    pub fn get_res_type(&self, i: usize) -> u8 {
         self.coords_vec[i].res_type
     }
 
     /// Provides the type of the `i`-th atom of this system
-    pub fn atom_type(&self, i: usize) -> u8 {
+    pub fn get_atom_type(&self, i: usize) -> u8 {
         self.coords_vec[i].atom_type
     }
 
@@ -232,7 +189,7 @@ impl Coordinates {
     /// by the `k`-th range: from `chain_ranges[k].0` to `chain_ranges[k].1`
     ///
     /// Assigning all the chains for this [`Coordinates`] help avoiding incorrect assignments
-    pub fn set_chains(&mut self, chain_ranges: &Vec<(usize, usize)>) {
+    pub fn set_chain_ranges(&mut self, chain_ranges: &Vec<(usize, usize)>) {
         assert_eq!(
             chain_ranges.last().unwrap().1,
             self.current_size,
@@ -259,17 +216,17 @@ impl Coordinates {
     }
 
     /// 'x' coordinate of `i`-th atom of this system
-    pub fn x(&self, i: usize) -> f64 {
+    pub fn get_x(&self, i: usize) -> f64 {
         self.coords_vec[i].x
     }
 
     /// 'y' coordinate of `i`-th atom of this system
-    pub fn y(&self, i: usize) -> f64 {
+    pub fn get_y(&self, i: usize) -> f64 {
         self.coords_vec[i].y
     }
 
     /// 'z' coordinate of `i`-th atom of this system
-    pub fn z(&self, i: usize) -> f64 {
+    pub fn get_z(&self, i: usize) -> f64 {
         self.coords_vec[i].z
     }
 
@@ -285,20 +242,20 @@ impl Coordinates {
         wrap_coordinate_to_box!(z, self.box_len, self.coords_vec[i].z);
     }
 
-    pub fn set(&mut self, i: usize, x: f64, y: f64, z: f64) {
+    pub fn set_xyz(&mut self, i: usize, x: f64, y: f64, z: f64) {
         wrap_coordinate_to_box!(x, self.box_len, self.coords_vec[i].x);
         wrap_coordinate_to_box!(y, self.box_len, self.coords_vec[i].y);
         wrap_coordinate_to_box!(z, self.box_len, self.coords_vec[i].z);
     }
 
-    pub fn set_vec(&mut self, i: usize, vec: Vec3) {
+    pub fn set_vec3(&mut self, i: usize, vec: Vec3) {
         let x = vec.x;
         let y = vec.y;
         let z = vec.z;
-        self.set(i, x, y, z);
+        self.set_xyz(i, x, y, z);
     }
 
-    pub fn add(&mut self, i: usize, x: f64, y: f64, z: f64) {
+    pub fn add_xyz(&mut self, i: usize, x: f64, y: f64, z: f64) {
         wrap_coordinate_to_box!(self.coords_vec[i].x + x, self.box_len, self.coords_vec[i].x);
         wrap_coordinate_to_box!(self.coords_vec[i].y + y, self.box_len, self.coords_vec[i].y);
         wrap_coordinate_to_box!(self.coords_vec[i].z + z, self.box_len, self.coords_vec[i].z);
@@ -308,7 +265,7 @@ impl Coordinates {
     ///
     /// This method (unlike the [`set()`](set) method) does not apply PBC. To the contrary,
     /// it assumes the two systems: `self` and `rhs` have exactly the same simulation box geometry
-    pub fn copy_from_vec(&mut self, i: usize, rhs: &Vec3) {
+    pub fn from_vec(&mut self, i: usize, rhs: &Vec3) {
         self.coords_vec[i].x = rhs.x;
         self.coords_vec[i].y = rhs.y;
         self.coords_vec[i].z = rhs.z;
@@ -330,7 +287,7 @@ impl IndexMut<usize> for Coordinates {
 
 impl System for Coordinates {
     /// Returns the current number of atoms of this system
-    fn size(&self) -> usize {
+    fn get_size(&self) -> usize {
         return self.current_size;
     }
 
@@ -352,7 +309,7 @@ impl ResizableSystem for Coordinates {
     }
 
     /// Returns the maximum number of atoms of this system
-    fn capacity(&self) -> usize {
+    fn get_capacity(&self) -> usize {
         return self.coords_vec.len();
     }
 }
