+++ bioshell-cartesians/src/coordinates_view.rs
@@ -1,0 +1,6 @@
+use crate::Coordinates;
+
+/// Stateless immutable view of coordinates
+pub struct CoordinatesView<'a> {
+    pub points: &'a Coordinates,
+}
