+++ bioshell-cartesians/src/movers/terminal_move.rs
@@ -4,7 +4,7 @@ use crate::CartesianSystem;
 use bioshell_montecarlo::{AcceptanceCriterion, AcceptanceStatistics, Mover};
 use bioshell_sim::{Energy, System};
 use bioshell_numerical::vec3::Vec3;
-use bioshell_numerical::{random_point_nearby, Rototranslation};
+use bioshell_numerical::{Rototranslation};
 
 
 /// A mover that applies a terminal move to a molecule.
@@ -37,7 +37,7 @@ impl<E: Energy<CartesianSystem>> Mover<CartesianSystem, E> for TerminalMove {
         acc: &mut dyn AcceptanceCriterion,
     ) -> Option<Range<usize>> {
         let mut rng = rand::thread_rng();
-        let size = system.size();
+        let size = system.get_size();
         let terminal_flag = rng.gen_range(0..1 + 1);
 
         let mut i_start = 0;
@@ -48,13 +48,13 @@ impl<E: Energy<CartesianSystem>> Mover<CartesianSystem, E> for TerminalMove {
         if terminal_flag == 0 {
             i_start = 0;
             i_end = self.frag_size;
-            terminal_vec = system.coordinates()[i_end-1];
-            term_rot_axis_start = system.coordinates()[i_end];
+            terminal_vec = system.get_coordinates()[i_end-1];
+            term_rot_axis_start = system.get_coordinates()[i_end];
         } else {
             i_start = size - self.frag_size;
             i_end = size;
-            term_rot_axis_start = system.coordinates()[i_start];
-            terminal_vec = system.coordinates()[i_start+1];
+            term_rot_axis_start = system.get_coordinates()[i_start];
+            terminal_vec = system.get_coordinates()[i_start+1];
         }
 
         let angle = rng.gen_range(-&self.max_angle..self.max_angle);
@@ -66,9 +66,9 @@ impl<E: Energy<CartesianSystem>> Mover<CartesianSystem, E> for TerminalMove {
 
         // --- apply forward rotation
         for i in i_start..i_end {
-            let mut temp_coord: Vec3 = system.coordinates()[i].clone();
+            let mut temp_coord: Vec3 = system.get_coordinates()[i].clone();
             roto_tran.apply_mut(&mut temp_coord);
-            system.set_vec(i, temp_coord);
+            system.set_vec3(i, temp_coord);
         }
 
         let energy_after = energy.energy(system);
@@ -80,9 +80,9 @@ impl<E: Energy<CartesianSystem>> Mover<CartesianSystem, E> for TerminalMove {
             self.succ_rate.n_failed += 1;
             // --- fall back - apply inverse rotation
             for i in i_start..i_end {
-                let mut temp_coord: Vec3 = system.coordinates()[i].clone();
+                let mut temp_coord: Vec3 = system.get_coordinates()[i].clone();
                 roto_tran.apply_inverse_mut(&mut temp_coord);
-                system.set_vec(i, temp_coord);
+                system.set_vec3(i, temp_coord);
             }
             return None;
         }
