+++ bioshell-cartesians/tests/test_coordinates.rs
@@ -9,9 +9,9 @@ mod coordinates_test {
     #[test]
     fn coordinates_initialization_test() {
         let mut coords = Coordinates::new(3);//create a coordinate of length: 3.
-        coords.add(0, 1.0,1.0,1.0);
-        coords.add(1, 2.0,2.0,2.0);
-        coords.add(2, 3.0,3.0,3.0);
+        coords.add_xyz(0, 1.0, 1.0, 1.0);
+        coords.add_xyz(1, 2.0, 2.0, 2.0);
+        coords.add_xyz(2, 3.0, 3.0, 3.0);
 
         assert_eq_vec3!(coords[0], Vec3::new(1.0, 1.0, 1.0), 1e-5);
         assert_eq_vec3!(coords[1], Vec3::new(2.0, 2.0, 2.0), 1e-5);
@@ -19,35 +19,35 @@ mod coordinates_test {
     }
 
     #[test]
-    fn clone_closest_image() {
+    fn clone_closest_image_test() {
         let mut coords = Coordinates::new(3);//create a coordinate of length: 3.
-        coords.add(0, 1.0,1.0,0.0);
-        coords.add(1, 3.0,2.0,1.0);
-        coords.add(2, 5.0,2.0,0.0);
+        coords.add_xyz(0, 1.0, 1.0, 0.0);
+        coords.add_xyz(1, 3.0, 2.0, 1.0);
+        coords.add_xyz(2, 5.0, 2.0, 0.0);
         coords.set_box_len(5.0);
 
-        let img = coords.clone_closest_image(0, 2);
+        let img = coords.get_closest_image_clone(0, 2);
         assert_eq_vec3!(img, Vec3::new(0.0, 2.0, 0.0), 1e-5);
 
-        coords.set(0, 0.0,1.0,1.0);
-        coords.set(1, 6.0,2.0,3.0);
-        coords.set(2, 0.0,2.0,5.0);
-        let img = coords.clone_closest_image(0, 2);
+        coords.set_xyz(0, 0.0, 1.0, 1.0);
+        coords.set_xyz(1, 6.0, 2.0, 3.0);
+        coords.set_xyz(2, 0.0, 2.0, 5.0);
+        let img = coords.get_closest_image_clone(0, 2);
         assert_eq_vec3!(img, Vec3::new(0.0, 2.0, 0.0), 1e-5);
     }
 
     #[test]
-    fn closest_distance() {
+    fn closest_distance_test() {
         let mut coords = Coordinates::new(3);//create a coordinate of length: 3.
-        coords.add(0, 1.0,1.0,0.0);
-        coords.add(1, 3.0,2.0,1.0);
-        coords.add(2, 5.0,2.0,0.0);
+        coords.add_xyz(0, 1.0, 1.0, 0.0);
+        coords.add_xyz(1, 3.0, 2.0, 1.0);
+        coords.add_xyz(2, 5.0, 2.0, 0.0);
         coords.set_box_len(5.0);
 
-        let d2 = coords.closest_distance_square(0, 2);
+        let d2 = coords.get_closest_distance_square(0, 2);
         assert_eq_float!(d2, 2.0, 1e-5);
 
-        let d2 = coords.closest_distance_square_to_vec(0, &Vec3::new(5.0, 2.0, 0.0));
+        let d2 = coords.get_closest_distance_square_to_vec(0, &Vec3::new(5.0, 2.0, 0.0));
         assert_eq_float!(d2, 2.0, 1e-5);
     }
 }
