+++ bioshell-ff/src/bonded.rs
@@ -17,7 +17,7 @@ impl SimpleHarmonic {
 
 macro_rules! spring_kernel {
     ($system:expr, $i:expr, $j:expr, $d0:expr, $en:expr, $OP:tt) => {
-        let v:f64 = ($system.closest_distance_square($i, $j).sqrt() - $d0) as f64;
+        let v:f64 = ($system.get_closest_distance_square($i, $j).sqrt() - $d0) as f64;
         $en $OP v*v;
     }
 }
@@ -25,9 +25,9 @@ macro_rules! spring_kernel {
 impl Energy<CartesianSystem> for SimpleHarmonic {
     fn energy(&self, system: &CartesianSystem) -> f64 {
         let mut en: f64 = 0.0;
-        for i in 1..system.size() {
-            if system.coordinates()[i].chain_id == system.coordinates()[i - 1].chain_id {
-                spring_kernel!(system.coordinates(),i, i-1, self.d0, en, +=);
+        for i in 1..system.get_size() {
+            if system.get_coordinates()[i].chain_id == system.get_coordinates()[i - 1].chain_id {
+                spring_kernel!(system.get_coordinates(),i, i-1, self.d0, en, +=);
             }
         }
 
@@ -36,13 +36,13 @@ impl Energy<CartesianSystem> for SimpleHarmonic {
 
     fn energy_by_pos(&self, system: &CartesianSystem, pos: usize) -> f64 {
         let mut en: f64 = 0.0;
-        if pos > 0 && system.coordinates()[pos].chain_id == system.coordinates()[pos - 1].chain_id {
-            spring_kernel!(system.coordinates(),pos, pos-1, self.d0, en, +=);
+        if pos > 0 && system.get_coordinates()[pos].chain_id == system.get_coordinates()[pos - 1].chain_id {
+            spring_kernel!(system.get_coordinates(),pos, pos-1, self.d0, en, +=);
         }
-        if pos < system.size() - 1
-            && system.coordinates()[pos].chain_id == system.coordinates()[pos + 1].chain_id
+        if pos < system.get_size() - 1
+            && system.get_coordinates()[pos].chain_id == system.get_coordinates()[pos + 1].chain_id
         {
-            spring_kernel!(system.coordinates(),pos, pos+1, self.d0, en, +=);
+            spring_kernel!(system.get_coordinates(),pos, pos+1, self.d0, en, +=);
         }
 
         return en * self.k as f64;
@@ -59,14 +59,14 @@ impl Energy<CartesianSystem> for SimpleHarmonic {
         } else {
             range.start
         };
-        let end: usize = if range.end >= system.size() - 1 {
-            system.size() - 1
+        let end: usize = if range.end >= system.get_size() - 1 {
+            system.get_size() - 1
         } else {
             range.end + 1
         };
         for ipos in start..end {
-            if system.coordinates()[ipos].chain_id == system.coordinates()[ipos + 1].chain_id {
-                spring_kernel!(system.coordinates(), ipos, ipos+1, self.d0, total_en, +=);
+            if system.get_coordinates()[ipos].chain_id == system.get_coordinates()[ipos + 1].chain_id {
+                spring_kernel!(system.get_coordinates(), ipos, ipos+1, self.d0, total_en, +=);
             }
         }
