+++ bioshell-montecarlo/src/lib.rs
@@ -1,3 +1,24 @@
-mod montecarlo;
 
-pub use montecarlo::*;
+// Import the classes from the separate files
+mod perm;
+mod acceptance_statistics;
+mod trait_acceptance_criterion;
+mod metropolis_criterion;
+mod trait_mover;
+mod trait_sampler;
+mod isothermal_mc;
+mod trait_stepwise_mover;
+mod trait_stepwise_builder;
+mod adaptive_mc_protocol;
+
+// Re-export the classes to make them accessible from outside the module
+pub use perm::*;
+pub use  acceptance_statistics::*;
+pub use  trait_acceptance_criterion::*;
+pub use  metropolis_criterion::*;
+pub use  trait_mover::*;
+pub use  trait_sampler::*;
+pub use  isothermal_mc::*;
+pub use  trait_stepwise_mover::*;
+pub use  trait_stepwise_builder::*;
+pub use  adaptive_mc_protocol::*;
