+++ bioshell-numerical/src/lib.rs
@@ -7,7 +7,7 @@ pub use matrix::*;
 pub use rototranslation::*;
 pub use vec3::*;
 
-mod testing_macros;
+mod macros;
 
-pub use testing_macros::*;
+pub use macros::*;
