+++ examples/polymer_crankshaft.rs
@@ -1,18 +1,16 @@
 use std::any::Any;
-use clap::builder::TypedValueParser;
 use std::env;
 use std::f32::consts::PI;
-use std::time::Instant;
 
 use clap::Parser;
 use log::info;
 
 use bioshell_cartesians::movers::{CrankshaftMove, TerminalMove};//import single-atom-mover
-use bioshell_cartesians::observers::{GyrationSquared, PdbTrajectory, REndSquared};
-use bioshell_cartesians::{
-    box_width, coordinates_to_pdb, pdb_to_coordinates, CartesianSystem, Coordinates, NbList,
-    PolymerRules, RandomChain,
-};
+use bioshell_cartesians::gyration_squared::{GyrationSquared};
+use bioshell_cartesians::pdb_trajectory::{PdbTrajectory};
+use bioshell_cartesians::r_end_squared::{REndSquared};
+use bioshell_cartesians::{coordinates_to_pdb, pdb_to_coordinates, CartesianSystem, Coordinates, NbList, PolymerRules, compute_box_width};
+use bioshell_cartesians::random_chain::RandomChain;
 use bioshell_core::utils::out_writer;
 
 use bioshell_sim::{Energy, Observer, ObserversSet, System};
@@ -95,7 +93,7 @@ pub fn main() {
     if args.infile == "" {//if 'infile' is empty,
         n_beads = args.n_beads;//set the number of beads
         coords = Coordinates::new(n_beads);//create `n_beads` number of coordinates
-        box_length = box_width(E_REP, n_beads, density);//calculate 3D box length
+        box_length = compute_box_width(E_REP, n_beads, density);//calculate 3D box length
         coords.set_box_len(box_length);//set box length to the calculated length
     }
     else
@@ -108,8 +106,8 @@ pub fn main() {
             }
             _ => panic!("Can't read from file >{}<", args.infile),
         };
-        n_beads = coords.size();//get number of atoms
-        box_length = box_width(E_REP, n_beads, density);//calculate box length
+        n_beads = coords.get_size();//get number of atoms
+        box_length = compute_box_width(E_REP, n_beads, density);//calculate box length
         coords.set_box_len(box_length);//set coords to box length
     }
 
@@ -127,28 +125,28 @@ pub fn main() {
 
     // ---------- Create a non-bonded list of neighbors
     let nbl: NbList = NbList::new(E_TO, buffer_thickness, Box::new(PolymerRules {}));
-    
-	// ---------- Create the system
+
+    // ---------- Create the system
     let mut system: CartesianSystem = CartesianSystem::new(coords, nbl);
-    
-	// ---------- Contact energy
+
+    // ---------- Contact energy
     let contact_kernel = SimpleContact::new(E_REP, E_FROM, E_TO, REP_VAL, E_VAL);
     let contacts: PairwiseNonbondedEvaluator<SimpleContact> =
         PairwiseNonbondedEvaluator::new(E_TO as f64, contact_kernel);
-    
-	// ---------- Harmonic energy (i.e. springs between beads)
+
+    // ---------- Harmonic energy (i.e. springs between beads)
     let harmonic = SimpleHarmonic::new(3.8, 1.0);
-    
-	// ---------- Total energy contains the contacts energy and bond springs
+
+    // ---------- Total energy contains the contacts energy and bond springs
     let mut total = TotalEnergy::new();
     total.add_component(Box::new(harmonic), 1.0);
     total.add_component(Box::new(contacts), 1.0);
-    if system.size() == 0 {
+    if system.get_size() == 0 {
         RandomChain::default().build(&mut system, &total);
     }
 
     // ---------- Show the starting point
-    coordinates_to_pdb(&system.coordinates(), 0, tra_f_name.as_str(), false);
+    coordinates_to_pdb(&system.get_coordinates(), 0, tra_f_name.as_str(), false);
     println!("# starting energy: {}", total.energy(&system));
 
     // ---------- Create a sampler and add a mover into it
@@ -178,7 +176,7 @@ pub fn main() {
     );
 
     coordinates_to_pdb(
-        &system.coordinates(),
+        &system.get_coordinates(),
         1i16,
         format!("{}final.pdb", &prefix).as_str(),
         false,
@@ -205,12 +203,12 @@ impl Observer for BondViolationObserver {
     type S = CartesianSystem;
 
     fn observe(&mut self, object: &Self::S) {
-        let coords = object.coordinates();
+        let coords = object.get_coordinates();
         let mut out_writer = out_writer(&self.out_fname, self.if_append);
         out_writer
             .write(format!("{:.6} ", self.i_model).as_bytes())
             .ok();
-        for ic in 0..coords.count_chains() {
+        for _ic in 0..coords.get_chains_count() {
             // HERE !!!
             let result: f64 = 0.12345;
             out_writer.write(format!("{:>10.3} ", result).as_bytes()).ok();
